[Bug 1781] longlong undefined in sntp handle_pkt() on Debian amd64.
Bump deps-ver and sntp/deps-ver to force each directory to be cleaned
once, to work around build breaks triggered by changing erealloc() and
estrdup() from functions to macros.

+++ sntp/main.c
@@ -205,7 +205,7 @@ handle_pkt (
 	char *p_SNTP_PRETEND_TIME;
 	time_t pretend_time;
 #if SIZEOF_TIME_T == 8
-	longlong ll;
+	long long ll;
 #else
 	long l;
 #endif
